requirements.txt
asgiref==3.6.0
backports.zoneinfo;python_version<"3.9"
Django==4.1.7
django-filter==22.1
django-phonenumber-field==7.0.2
djangorestframework==3.14.0
drf-nested-routers==0.93.4
importlib-metadata==6.1.0
Markdown==3.4.1
phonenumbers==8.13.7
Pillow==9.4.0
pytz==2022.7.1
sqlparse==0.4.3
zipp==3.15.0
djangorestframework-simplejwt
